Time: 0 ms (100%), Space: 9.7 MB (80.24%) - LeetHub

+++ 1635-number-of-good-pairs/1635-number-of-good-pairs.cpp
@@ -1,9 +1,9 @@
 class Solution {
 public:
     int numIdenticalPairs(vector<int>& nums) {
-        int count = 0;
-        for (int i = 0; i < size(nums); i++) {
-            for (int j = 0; j < size(nums); j++) {
+        size_t count = 0;
+        for (size_t i = 0; i < size(nums); i++) {
+            for (size_t j = 0; j < size(nums); j++) {
                 if (nums[i] == nums[j] && i < j) {
                     count++;
                 }
